requirements.txt
feedparser
anthropic
pyyaml
python-dotenv
requests
beautifulsoup4
lxml
